Player animations based on speed fixed bug where enemies do not spawn during course of level A new level Info system to adjust level time and type I have the one hit kill system done, meaning the health UI is gone and the level finished / you died screen works **along with the start of the new core loop**
You enter into the game: Instead of there being a timer counting down, we have a min goal time. Once the timer has passed this time, then we can see the dev goal time. Once the player has passed the min time for a level, the door appears. Currently, once the door appears the player can enter into it when they want or keep going to try to beat the dev time. I am also wondering if a system for implementing a spawn system for when the player is trying for the dev time. **If the player dies before the min time, they wont be able to play the next level. If they die after the min time, then they have passed the level and go.**

TODO:
Level selection / map system
Saved time for each level
Death Counter
meshing all of the above together